testutil: Rename func from Run To RunCRUD and improve some
