+++ v2/helper/builder/archive/standard_archive_builder_test.go
@@ -28,7 +28,7 @@ func TestStandardArchiveBuilder_Build(t *testing.T) {
 	testZone := testutil.TestZone()
 	var sourceArchive *sacloud.Archive
 
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		SetupAPICallerFunc: func() sacloud.APICaller {
 			return testutil.SingletonAPICaller()
 		},
