+++ v2/helper/builder/internet/builder_test.go
@@ -26,7 +26,7 @@ import (
 func TestBuilder_Build(t *testing.T) {
 	var testZone = testutil.TestZone()
 
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		SetupAPICallerFunc: func() sacloud.APICaller {
 			return testutil.SingletonAPICaller()
 		},
