+++ v2/helper/builder/localrouter/builder_test.go
@@ -27,7 +27,7 @@ func TestBuilder_Build(t *testing.T) {
 	var peerLocalRouter *sacloud.LocalRouter
 	var sw *sacloud.Switch
 
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		SetupAPICallerFunc: func() sacloud.APICaller {
 			return testutil.SingletonAPICaller()
 		},
