+++ v2/helper/builder/registry/builder_test.go
@@ -23,7 +23,7 @@ import (
 )
 
 func TestBuilder_Build(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		SetupAPICallerFunc: func() sacloud.APICaller {
 			return testutil.SingletonAPICaller()
 		},
