+++ v2/sacloud/test/auto_backup_op_test.go
@@ -25,7 +25,7 @@ import (
 )
 
 func TestAutoBackupOpCRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel: true,
 
 		SetupAPICallerFunc: singletonAPICaller,
