+++ v2/sacloud/test/bridge_op_test.go
@@ -22,7 +22,7 @@ import (
 )
 
 func TestBridgeOpCRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel: true,
 
 		SetupAPICallerFunc: singletonAPICaller,
