+++ v2/sacloud/test/container_registry_op_test.go
@@ -23,7 +23,7 @@ import (
 )
 
 func TestContainerRegistryOp_CRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel:           true,
 		SetupAPICallerFunc: singletonAPICaller,
 		Create: &testutil.CRUDTestFunc{
