+++ v2/sacloud/test/database_op_test.go
@@ -25,7 +25,7 @@ import (
 )
 
 func TestDatabaseOpCRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel: true,
 
 		SetupAPICallerFunc: singletonAPICaller,
