+++ v2/sacloud/test/dns_op_test.go
@@ -23,7 +23,7 @@ import (
 )
 
 func TestDNSOp_CRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel: true,
 
 		SetupAPICallerFunc: singletonAPICaller,
