+++ v2/sacloud/test/gslb_op_test.go
@@ -23,7 +23,7 @@ import (
 )
 
 func TestGSLBOp_CRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel: true,
 
 		SetupAPICallerFunc: singletonAPICaller,
