+++ v2/sacloud/test/license_op_test.go
@@ -23,7 +23,7 @@ import (
 )
 
 func TestLicenseOpCRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel:           true,
 		IgnoreStartupWait:  true,
 		SetupAPICallerFunc: singletonAPICaller,
