+++ v2/sacloud/test/load_balancer_op_test.go
@@ -24,7 +24,7 @@ import (
 )
 
 func TestLoadBalancerOp_CRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel: true,
 
 		SetupAPICallerFunc: singletonAPICaller,
