+++ v2/sacloud/test/nfs_op_test.go
@@ -25,7 +25,7 @@ import (
 )
 
 func TestNFSOp_CRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel: true,
 
 		SetupAPICallerFunc: singletonAPICaller,
