+++ v2/sacloud/test/packet_filter_op_test.go
@@ -23,7 +23,7 @@ import (
 )
 
 func TestPacketFilterOp_CRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel:           true,
 		IgnoreStartupWait:  true,
 		SetupAPICallerFunc: singletonAPICaller,
