+++ v2/sacloud/test/private_host_op_test.go
@@ -22,7 +22,7 @@ import (
 )
 
 func TestPrivateHostOp_CRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel:           true,
 		IgnoreStartupWait:  true,
 		SetupAPICallerFunc: singletonAPICaller,
