+++ v2/sacloud/test/vpc_router_op_test.go
@@ -25,7 +25,7 @@ import (
 )
 
 func TestVPCRouterOp_CRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel:           true,
 		SetupAPICallerFunc: singletonAPICaller,
 		Create: &testutil.CRUDTestFunc{
@@ -170,7 +170,7 @@ func testVPCRouterDelete(ctx *testutil.CRUDTestContext, caller sacloud.APICaller
 }
 
 func TestVPCRouterOp_WithRouterCRUD(t *testing.T) {
-	testutil.Run(t, &testutil.CRUDTestCase{
+	testutil.RunCRUD(t, &testutil.CRUDTestCase{
 		Parallel:           true,
 		SetupAPICallerFunc: singletonAPICaller,
